backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML and embeddings
sentence-transformers==2.3.1
chromadb==0.4.22

# HTTP client for OpenRouter
httpx==0.26.0

# Rate limiting
slowapi==0.1.9

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
